Reworked example/command structure.
Needed a way to get iBoss to communicate with exact properly without having to write 100's of lines of compat code.
Easiest way was to just turn the stuff into a simple container based application during dev.
This allows for seamless creation between iBoss primary container and this small dev container.

This dev container is nothing more than a glorified constructor.
But what it does do, it does properly.