End of day commit.
Posting new stuff has been failing.
Documentation on this is extensive on HOW to do it, but WHERE to put the data we want.
Using old AF commits from iBoss to create a test flow right now.
Looking solid so far.

May need to invest in a cached version of findOneBy/matching.
This is eating through the quota limits like it's nothing.
A single invoice would easily eat 10+ calls.
That is not sustainable when we're doing it for more than 1 client.